:truck: Separate client and server
