+++ client/src/app/search-bar/search-bar.component.spec.ts
