+++ client/src/app/search-bar/search-bar.component.ts
