+++ client/src/app/weather/weather.component.ts
@@ -1,8 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { response } from 'express';
-import * as dotenv from 'dotenv';
-dotenv.config();
+
 const KEY = process.env.KEY || '';
 
 @Component({
@@ -12,15 +10,16 @@ const KEY = process.env.KEY || '';
 })
 export class WeatherComponent implements OnInit {
   lat: string;
-  long :string;
-  weather : object =null;
+  long: string;
+  weather: object = null;
   constructor(private http: HttpClient) { }
 
   ngOnInit() {
   }
   getWeather(lat: string, long: string) {
-    console.log("Weather componenet:",lat,",",long);
-    this.http.get('https://api.darksky.net/forecast/'+ KEY + '/'+lat+','+long)
-    .subscribe( this.weather = response.get);
+    console.log("Weather componenet:", lat, ",", long);
+    this.http.get('https://api.darksky.net/forecast/' + KEY + '/' + lat + ',' + long)
+    // FIXME: subscribe takes a function - see the RxJS lecture on Panopto for reference
+      .subscribe(this.weather = response.get);
   }
 }
