+++ client/src/favicon.ico
