+++ client/src/karma.conf.js
