+++ client/src/tsconfig.app.json
